Added new functionality to allow users to send content to be diffed as parameters to the 'prettyTextDiff' function.
Currently, the user has to set th text of the HTML elements i.e. the originalContainer and the changedContainer. This required the HTML content of the two containers to be part of the page source which caused web pages to be prone to XSS attacks. This method allows the user to set the HTML content to a JS variable and pass it as a parameter to this function. Both functionalities are supported to not break the existing plugin.